Update main.c
In previous file we cant get accurate date, it was a wrong and it is configured for leap year. without change in number of days array, but now i have refined the code accordingly for both leap and non-leap year, please have a look at it. If can be optimized in different way, let me know.